requirements.txt
streamlit
pymupdf
langchain
langchain-community
langchain-text-splitters
sentence-transformers
chromadb
groq
python-dotenv